+++ src/actions/deviceActions.ts
@@ -64,18 +64,25 @@ import {
 import { updateGainsAction } from '../slices/gainsSlice';
 import {
     clearProgress,
+    DigitalChannelTriggerLogic,
+    DigitalChannelTriggerStatesEnum,
+    getTriggerOffset,
     getTriggerRecordingLength,
     resetTriggerOrigin,
     setProgress,
     setTriggerActive,
     setTriggerOrigin,
+    TriggerEdge,
 } from '../slices/triggerSlice';
 import { updateRegulator as updateRegulatorAction } from '../slices/voltageRegulatorSlice';
 import { convertBits16 } from '../utils/bitConversion';
 import { convertTimeToSeconds } from '../utils/duration';
 import { isDiskFull } from '../utils/fileUtils';
 import { isDataLoggerPane } from '../utils/panes';
-import { setSpikeFilter as persistSpikeFilter } from '../utils/persistentStore';
+import {
+    digitalChannelStateTupleOf8,
+    setSpikeFilter as persistSpikeFilter,
+} from '../utils/persistentStore';
 
 let device: null | SerialDevice = null;
 let updateRequestInterval: NodeJS.Timeout | undefined;
@@ -212,6 +219,68 @@ const initGains = (): AppThunk<RootState, Promise<void>> => async dispatch => {
     );
 };
 
+function checkDigitalTriggerValidity(
+    unsignedBits: number,
+    previousUnsignedBits: number,
+    channelTriggerStatuses: digitalChannelStateTupleOf8,
+    digitalTriggerLogic: DigitalChannelTriggerLogic
+): boolean {
+    if (unsignedBits === previousUnsignedBits) {
+        return false;
+    }
+
+    const result = channelTriggerStatuses
+        .map((state, index) => {
+            const prevBit = getBit(previousUnsignedBits, index);
+            const currBit = getBit(unsignedBits, index);
+
+            if (state === DigitalChannelTriggerStatesEnum.High)
+                return currBit === 1;
+            if (state === DigitalChannelTriggerStatesEnum.Low)
+                return currBit === 0;
+            if (state === DigitalChannelTriggerStatesEnum.Any)
+                return prevBit !== currBit;
+            return null;
+        })
+        .filter(r => r !== null);
+
+    switch (digitalTriggerLogic) {
+        case 'AND':
+            return result.every(Boolean);
+        case 'OR':
+            return result.some(Boolean);
+        default:
+            return false;
+    }
+}
+
+const getBit = (number: number, position: number) => (number >> position) & 1;
+
+function checkAnalogTriggerValidity(
+    cappedValue: number,
+    prevCappedValue: number | undefined,
+    triggerLevel: number,
+    triggerEdge: TriggerEdge
+): boolean {
+    const isRisingEdge = triggerEdge === 'Rising Edge';
+    const isLoweringEdge = triggerEdge === 'Falling Edge';
+
+    let validTriggerValue = false;
+
+    if (isRisingEdge) {
+        validTriggerValue =
+            prevCappedValue != null &&
+            prevCappedValue < triggerLevel &&
+            cappedValue >= triggerLevel;
+    } else if (isLoweringEdge) {
+        validTriggerValue =
+            prevCappedValue != null &&
+            prevCappedValue > triggerLevel &&
+            cappedValue <= triggerLevel;
+    }
+    return validTriggerValue;
+}
+
 export const open =
     (deviceInfo: Device): AppThunk<RootState, Promise<void>> =>
     async (dispatch, getState) => {
@@ -225,15 +294,17 @@ export const open =
         let prevValue = 0;
         let prevCappedValue: number | undefined;
         let prevBits = 0;
+        let prevUnsignedBits = 0;
         let nbSamples = 0;
         let nbSamplesTotal = 0;
 
         const onSample = ({ value, bits }: SampleValues) => {
+            const state = getState();
             const {
                 app: { samplingRunning },
                 dataLogger: { maxSampleFreq },
-            } = getState().app;
-            const sampleFreq = getSampleFrequency(getState());
+            } = state.app;
+            const sampleFreq = getSampleFrequency(state);
             if (!samplingRunning) {
                 return;
             }
@@ -244,6 +315,9 @@ export const open =
                 cappedValue = 0;
             }
 
+            const channelTriggerStatuses =
+                state.app.trigger.digitalChannelsTriggersStates;
+            const unsignedBits = bits !== undefined ? bits & 0xff : 0;
             const b16 = convertBits16(bits!);
 
             if (samplingRunning && sampleFreq < maxSampleFreq) {
@@ -267,30 +341,50 @@ export const open =
             DataManager().addData(cappedValue, b16 | prevBits);
             prevBits = 0;
 
-            if (getRecordingMode(getState()) === 'Scope') {
+            if (getRecordingMode(state) === 'Scope') {
+                const triggerCategory = state.app.trigger.category;
+                const digitalTriggerLogic =
+                    state.app.trigger.digitalChannelsTriggerLogic;
+
                 const validTriggerValue =
-                    prevCappedValue != null &&
-                    prevCappedValue < getState().app.trigger.level &&
-                    cappedValue >= getState().app.trigger.level;
+                    triggerCategory === 'Analog'
+                        ? checkAnalogTriggerValidity(
+                              cappedValue,
+                              prevCappedValue,
+                              state.app.trigger.level,
+                              state.app.trigger.edge
+                          )
+                        : checkDigitalTriggerValidity(
+                              unsignedBits,
+                              prevUnsignedBits,
+                              channelTriggerStatuses,
+                              digitalTriggerLogic
+                          );
+
                 prevCappedValue = cappedValue;
+                prevUnsignedBits = unsignedBits;
 
                 if (!DataManager().isInSync()) {
                     return;
                 }
 
-                if (!getState().app.trigger.active && validTriggerValue) {
+                if (!state.app.trigger.active && validTriggerValue) {
                     if (latestTrigger !== undefined) {
                         return;
                     }
 
-                    if (!isSavePending(getState())) {
+                    if (!isSavePending(state)) {
                         dispatch(setSavePending(true));
                     }
                     dispatch(setTriggerActive(true));
+                    const offsetLength = getTriggerOffset(getState());
+                    const totalRecordingLength =
+                        getTriggerRecordingLength(getState()) + offsetLength;
                     dispatch(
                         processTrigger(
                             cappedValue,
-                            getTriggerRecordingLength(getState()) * 1000, // ms to uS
+                            totalRecordingLength * 1000, // ms to uS
+                            offsetLength * 1000,
                             (progressMessage, prog) => {
                                 dispatch(
                                     setProgress({
@@ -309,26 +403,26 @@ export const open =
                         }
                         if (
                             samplingRunning &&
-                            getState().app.trigger.type === 'Single'
+                            state.app.trigger.type === 'Single'
                         ) {
                             dispatch(samplingStop());
                         }
                     });
                 } else if (
-                    getState().app.trigger.active &&
+                    state.app.trigger.active &&
                     !validTriggerValue &&
-                    getState().app.trigger.type === 'Continuous'
+                    state.app.trigger.type === 'Continuous'
                 ) {
                     dispatch(setTriggerActive(false));
                 }
-            } else if (!isSavePending(getState())) {
+            } else if (!isSavePending(state)) {
                 dispatch(setSavePending(true));
             }
 
             const durationInMicroSeconds =
                 convertTimeToSeconds(
-                    getState().app.dataLogger.duration,
-                    getState().app.dataLogger.durationUnit
+                    state.app.dataLogger.duration,
+                    state.app.dataLogger.durationUnit
                 ) * microSecondsPerSecond;
             if (durationInMicroSeconds <= DataManager().getTimestamp()) {
                 if (samplingRunning) {
@@ -500,10 +594,14 @@ export const processTrigger =
     (
         triggerValue: number,
         triggerLength: number,
+        offsetLength: number,
         onProgress?: (message: string, progress?: number) => void
     ): AppThunk<RootState, Promise<void>> =>
     async (dispatch, getState) => {
-        const trigger = DataManager().addTimeReachedTrigger(triggerLength);
+        const trigger = DataManager().addTimeReachedTrigger(
+            triggerLength,
+            offsetLength
+        );
 
         const triggerTime = Date.now();
         const remainingRecordingLength = triggerLength / 2;