+++ src/components/SidePanel/StartStop.tsx
@@ -8,7 +8,6 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
     ConfirmationDialog,
-    Group,
     logger,
     StartStopButton,
     telemetry,
@@ -34,7 +33,17 @@ import {
     dataLoggerState,
     getSampleFrequency,
 } from '../../slices/dataLoggerSlice';
-import { resetTriggerOrigin } from '../../slices/triggerSlice';
+import {
+    getDigitalChannelsTriggerLogic,
+    getDigitalChannelsTriggersStates,
+    getTriggerCategory,
+    getTriggerEdge,
+    getTriggerOffset,
+    getTriggerRecordingLength,
+    getTriggerType,
+    getTriggerValue,
+    resetTriggerOrigin,
+} from '../../slices/triggerSlice';
 import { convertTimeToSeconds, formatDuration } from '../../utils/duration';
 import {
     calcFileSize,
@@ -47,8 +56,6 @@ import {
     setDoNotAskStartAndClear,
 } from '../../utils/persistentStore';
 import { resetCache } from '../Chart/data/dataAccumulator';
-import LiveModeSettings from './LiveModeSettings';
-import TriggerSettings from './TriggerSettings';
 
 const fmtOpts = { notation: 'fixed' as const, precision: 1 };
 
@@ -70,6 +77,13 @@ export default () => {
     const savePending = useSelector(isSavePending);
     const sessionFolder = useSelector(getSessionRootFolder);
     const diskFullTrigger = useSelector(getDiskFullTrigger);
+    const triggerCategory = useSelector(getTriggerCategory);
+    const triggerRecordingLength = useSelector(getTriggerRecordingLength);
+    const triggerOffset = useSelector(getTriggerOffset);
+    const triggerEdge = useSelector(getTriggerEdge);
+    const triggerLogic = useSelector(getDigitalChannelsTriggerLogic);
+    const triggerStates = useSelector(getDigitalChannelsTriggersStates);
+    const triggerValue = useSelector(getTriggerValue);
 
     const sampleIndefinitely = durationUnit === 'inf';
 
@@ -90,10 +104,28 @@ export default () => {
 
         const mode: RecordingMode = scopePane ? 'Scope' : 'DataLogger';
 
-        telemetry.sendEvent('StartSampling', {
+        const telemetryMetadata = {
             mode,
-            samplesPerSecond: DataManager().getSamplesPerSecond(),
-        });
+            ...(mode === 'DataLogger' && {
+                samplesPerSecond: DataManager().getSamplesPerSecond(),
+            }),
+            ...(mode === 'Scope' && {
+                triggerRecordingLength,
+                triggerOffset,
+                triggerCategory,
+                triggerType,
+                ...(triggerCategory === 'Analog' && {
+                    triggerValue,
+                    triggerEdge,
+                }),
+                ...(triggerCategory === 'Digital' && {
+                    triggerLogic,
+                    triggerStates,
+                }),
+            }),
+        };
+
+        telemetry.sendEvent('StartSampling', telemetryMetadata);
 
         if (mode === 'DataLogger') {
             if (!fs.existsSync(sessionFolder)) {
@@ -153,12 +185,10 @@ export default () => {
         setRemainingTime(calcRemainingTime(freeSpace, sampleFreq));
     }, [freeSpace, sampleFreq]);
 
+    const triggerType = useSelector(getTriggerType);
+
     return (
         <>
-            <Group heading="Sampling parameters" gap={4}>
-                {dataLoggerPane && <LiveModeSettings />}
-                {scopePane && <TriggerSettings />}
-            </Group>
             <div className="tw-flex tw-flex-col tw-gap-2">
                 <StartStopButton
                     title={startStopTitle}