+++ src/components/SidePanel/TriggerSettings/AnalogTriggerSettings.tsx
@@ -12,17 +12,15 @@ import {
     StateSelector,
 } from '@nordicsemiconductor/pc-nrfconnect-shared';
 
-import { appState } from '../../slices/appSlice';
-import { getRecordingMode } from '../../slices/chartSlice';
+import { appState } from '../../../slices/appSlice';
+import { getRecordingMode } from '../../../slices/chartSlice';
 import {
-    getTriggerRecordingLength,
-    getTriggerType,
+    getTriggerEdge,
     getTriggerValue,
+    setTriggerEdge,
     setTriggerLevel,
-    setTriggerRecordingLength,
-    setTriggerType,
-    TriggerTypeValues,
-} from '../../slices/triggerSlice';
+    TriggerEdgeValues,
+} from '../../../slices/triggerSlice';
 
 const CurrentUnitValues = ['mA', '\u00B5A'] as const;
 type CurrentUnit = (typeof CurrentUnitValues)[number];
@@ -47,9 +45,8 @@ const convertToMicroAmps = (unit: CurrentUnit, value: number) => {
 
 export default () => {
     const dispatch = useDispatch();
-    const recordingLength = useSelector(getTriggerRecordingLength);
     const triggerValue = useSelector(getTriggerValue);
-    const triggerType = useSelector(getTriggerType);
+    const triggerEdge = useSelector(getTriggerEdge);
     const { samplingRunning } = useSelector(appState);
     const dataLoggerActive =
         useSelector(getRecordingMode) === 'DataLogger' && samplingRunning;
@@ -63,8 +60,6 @@ export default () => {
 
     const [internalTriggerValue, setInternalTriggerValue] =
         useState(triggerValue);
-    const [internalTriggerLength, setInternalTriggerLength] =
-        useState(recordingLength);
 
     useEffect(() => {
         if (triggerValue > 1000) {
@@ -76,30 +71,8 @@ export default () => {
         }
     }, [triggerValue]);
 
-    useEffect(() => {
-        setInternalTriggerLength(recordingLength);
-    }, [recordingLength]);
-
     return (
         <>
-            <NumberInput
-                range={{
-                    min: 1,
-                    max: 1000,
-                    decimals: 2,
-                    step: 0.01,
-                }}
-                title="Duration of trigger window"
-                value={internalTriggerLength}
-                onChange={setInternalTriggerLength}
-                onChangeComplete={(value: number) => {
-                    dispatch(setTriggerRecordingLength(value));
-                }}
-                unit="ms"
-                label="Length"
-                disabled={dataLoggerActive}
-                showSlider
-            />
             <NumberInput
                 range={{
                     min: getMin(levelUnit),
@@ -134,10 +107,13 @@ export default () => {
                 showSlider
             />
             <StateSelector
-                items={[...TriggerTypeValues]}
-                onSelect={m => dispatch(setTriggerType(TriggerTypeValues[m]))}
-                selectedItem={triggerType}
+                items={[...TriggerEdgeValues]}
+                onSelect={m => {
+                    dispatch(setTriggerEdge(TriggerEdgeValues[m]));
+                }}
+                selectedItem={triggerEdge}
                 disabled={samplingRunning}
+                size="sm"
             />
         </>
     );