+++ src/globals.ts
@@ -366,7 +366,10 @@ export const DataManager = () => ({
             foldingBuffer: options.foldingBuffer,
         };
     },
-    addTimeReachedTrigger: (recordingLengthMicroSeconds: number) =>
+    addTimeReachedTrigger: (
+        recordingLengthMicroSeconds: number,
+        offsetLengthMicroSeconds: number
+    ) =>
         new Promise<{
             writeBuffer: WriteBuffer;
             timeRange: Range;
@@ -383,17 +386,13 @@ export const DataManager = () => ({
                 options.writeBuffer.getBytesWritten() / frameSize - 1;
             const timestamp = indexToTimestamp(currentIndex);
 
-            const splitRecordingLengthMicroSeconds =
-                recordingLengthMicroSeconds / 2;
+            const afterTriggerLength =
+                recordingLengthMicroSeconds - offsetLengthMicroSeconds;
+
+            // const splitRecordingLengthMicroSeconds = recordingLengthMicroSeconds / 2;
             const timeRange = {
-                start: Math.max(
-                    0,
-                    timestamp - splitRecordingLengthMicroSeconds
-                ),
-                end:
-                    timestamp +
-                    splitRecordingLengthMicroSeconds -
-                    indexToTimestamp(1), // we must exclude current sample the one that triggered all this
+                start: Math.max(0, timestamp - offsetLengthMicroSeconds),
+                end: timestamp + afterTriggerLength - indexToTimestamp(1), // we must exclude current sample the one that triggered all this
             };
 
             const bytesRange = {